+++ CHANGELOG.md
@@ -10,6 +10,17 @@ and this project adheres to
 
 ### Added
 
+- Create relationships for every member of `google_iam_binding`s.
+- Added support for ingesting the following **new** relationships:
+
+| Source                             | class      | Target                             |
+| ---------------------------------- | ---------- | ---------------------------------- |
+| `google_iam_binding`               | `ASSIGNED` | `google_cloud_authenticated_users` |
+| `google_iam_binding`               | `ASSIGNED` | `everyone`                         |
+| `google_iam_binding`               | `ASSIGNED` | `google_iam_role`                  |
+| `everyone`                         | `ASSIGNED` | `google_iam_role`                  |
+| `google_cloud_authenticated_users` | `ASSIGNED` | `google_iam_role`                  |
+
 - Added support for ingesting the following **new** resources:
 
   | Service        | Resource / Entity                                                                                                                       |